# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "0.0.1"
    },
    "paths": {
        "/jobs": {
            "get": {
                "operationId": "App\\Http\\Controllers\\JobController::index",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "operationId": "App\\Http\\Controllers\\JobController::store",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/jobs/{id}": {
            "get": {
                "operationId": "App\\Http\\Controllers\\JobController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Job id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "put": {
                "operationId": "App\\Http\\Controllers\\JobController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Job id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/job3s/{id}": {
            "delete": {
                "operationId": "App\\Http\\Controllers\\JobController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Job id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        }
    }
}